refactor: move cli command to dedicated modules
